event name fixed

+++ server/src/tickets/entity.ts
@@ -21,7 +21,7 @@ export default class Ticket extends BaseEntity {
   @Column('text', {nullable:true})
   picture: string
 
-  @Column('float', {default:0})
+  @Column('integer', {default:0})
   fraudRisk: number
 
   @CreateDateColumn({type: "timestamp"})
